Fixed bug where desired signal is below center frequency

+++ qpsk_detect.py
@@ -48,8 +48,8 @@ bandpass = args.bandpass
 symbol_rates = [72e3, 2e6, 4e6, 6e6] # Possible symbol rates
 
 with open(args.filename,'rb') as fp:
-    sample_rate = np.fromfile(fp, dtype='uint32', count=1, sep='')[0]
-    center_freq = np.fromfile(fp, dtype='uint64', count=1, sep='')[0]
+    sample_rate = np.fromfile(fp, dtype='uint32', count=1, sep='')[0].astype('float64')
+    center_freq = np.fromfile(fp, dtype='uint64', count=1, sep='')[0].astype('float64')
     timestamp = np.fromfile(fp, dtype='uint64', count=1, sep='')[0]
     sample_size = np.fromfile(fp, dtype='uint32', count=1, sep='')[0]
     crc = np.fromfile(fp, dtype='uint32', count=1, sep='',offset=4)[0]
