+++ qpsk_dopplerscan.py
@@ -136,8 +136,8 @@ device = args.device
 doppler_axis = np.linspace(dopplerstart,dopplerstop,dopplersamples)
 
 with open(args.filename,'rb') as fp:
-    sample_rate = np.fromfile(fp, dtype='uint32', count=1, sep='')[0]
-    center_freq = np.fromfile(fp, dtype='uint64', count=1, sep='')[0]
+    sample_rate = np.fromfile(fp, dtype='uint32', count=1, sep='')[0].astype('float64')
+    center_freq = np.fromfile(fp, dtype='uint64', count=1, sep='')[0].astype('float64')
     timestamp = np.fromfile(fp, dtype='uint64', count=1, sep='')[0]
     sample_size = np.fromfile(fp, dtype='uint32', count=1, sep='')[0]
     crc = np.fromfile(fp, dtype='uint32', count=1, sep='',offset=4)[0]
@@ -170,7 +170,7 @@ with open(args.filename,'rb') as fp:
         filt=torch.conj(rrcosfilter(window_size, 0.35, 1.0/bandpass, sample_rate).to(device))
     else:
         filt = None
-    
+
     for batch in tqdm.tqdm(range(batches),desc='Running batches',position=0):
         # Pull data from the file
         data = np.fromfile(fp,
